pyproject.toml
[tool.poetry]
name = "parser_ozon"
description = "Parser Ozon"
version = "0.1.0"
readme = "README.md"
authors = ["Pavel Korchagin <xoste49@gmail.com>"]

[tool.poetry.dependencies]
python = "3.10.12"

django = {extras = ["argon2"], version = "^4.2.3"}
django-split-settings = "^1.2.0"
django-axes = "^6.0.5"
django-health-check = "^3.17.0"

gunicorn = "^21.2.0"
python-decouple = "^3.8"
structlog = "^23.1.0"
psycopg = {extras = ["binary"], version = "^3.1.9"}
djangorestframework = "^3.14.0"
beautifulsoup4 = "^4.12.2"
requests = "^2.31.0"
selenium = "^4.10.0"
lxml = "^4.9.3"
undetected-chromedriver = "^3.5.0"
selenium-stealth = "^1.0.6"
celery = {extras = ["redis"], version = "^5.3.1"}
django-celery-results = "^2.5.1"
django-celery-beat = "^2.5.0"
aiogram = "^2.25.1"
drf-yasg = "^1.21.7"
django-adminlte3 = "^0.1.6"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.1"
django-querycount = "^0.8"
django-migration-linter = "^5.0"
django-extra-checks = "^0.13"
nplusone = "^1.0"

flake8-pytest-style = "^1.7"
flake8-django = "^1.3"

pytest = "^7.4"
pytest-django = "^4.5"
pytest-cov = "^4.0"
covdefaults = "^2.3"
django-coverage-plugin = "^3.1"
pytest-randomly = "^3.13"
pytest-timeout = "^2.1"
django-test-migrations = "^1.3"
hypothesis = "^6.82"

mypy = "^1.4"
django-stubs = "^4.2"

djlint = "^1.32"
yamllint = "^1.32"
safety = "^2.3"
dotenv-linter = "^0.4"
ipython = "^8.14"

[build-system]
requires = ["poetry-core>=1.6"]
build-backend = "poetry.core.masonry.api"


[tool.djlint]
ignore = "H006,H030,H031"
include = "H017,H035"
indent = 2
blank_line_after_tag = "load,extends"
profile = "django"
max_line_length = 80
format_attribute_template_tags = true

